refactor: create 'utils/k8s' package
